requirements.txt
# Social Listening Service Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# HTTP clients
httpx>=0.26.0
aiohttp>=3.9.0

# Supabase
supabase>=2.13.0

# Google Gemini AI (embeddings + classification)
google-generativeai>=0.8.0

# Google GenAI SDK v1.x (blog pipeline stages)
google-genai>=1.0.0

# Apify client (for platform crawlers)
apify-client>=1.0.0

# Environment
python-dotenv>=1.0.0

# HTML parsing
beautifulsoup4>=4.12.0

# Markdown processing
markdown>=3.5.0
